refractor: seperate constants into their own file
